% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot_day.R
\name{add_plot_day}
\alias{add_plot_day}
\title{Add Collapsed Integer Days For Plotting}
\usage{
add_plot_day(data)
}
\arguments{
\item{data}{data frame of results; must contain \code{season} and \code{date}
columns}
}
\description{
Adds an integer column \code{.plot_day} to a data frame of results that
represents a sequential time series, but with summer gaps removed.
Any results that actually occur during summer (Southern Hemisphere) will
be \code{NA}. Also adds attributes to the data frame called
\code{x_breaks} and \code{x_labels} that can be used when plotting to mark
the first of each year (Jan 1st).
}
