% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ath_plot.R
\name{ath_plot_dst}
\alias{ath_plot_dst}
\title{Distance Results Plots}
\usage{
ath_plot_dst(ath_names, races = c("maj_int", "fis"), by_tech = FALSE,
  collapse = NULL)
}
\arguments{
\item{ath_names}{character vector of athlete names}

\item{races}{character; either "maj_int" or "fis"}

\item{by_tech}{boolean; summarise by season or by season and technique}

\item{collapse}{names list of athletes to collapse into a single panel}
}
\description{
Distance Results Plots
}

