% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_funs.R
\name{bs_mads}
\alias{bs_mads}
\title{Bootstrap MAD}
\usage{
bs_mads(vec, B = 1000)
}
\arguments{
\item{vec}{vector of values}

\item{B}{integer; number of bootstrap replicates}
}
\value{
A list with components \code{orig} with the original statistic and
\code{bs} a matrix whose columns are the bootstrapped statistics
}
\description{
Bootstrap median absolute deviation of a vector.
}
