% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_funs.R
\name{bs_median}
\alias{bs_median}
\title{Bootstrap Median}
\usage{
bs_median(vec, B = 1000)
}
\arguments{
\item{vec}{vector of values}

\item{B}{integer; number of bootstrap replicates}
}
\value{
A list with components \code{orig} with the original statistic and
\code{bs} a vector of the bootstrapped statistics
}
\description{
Bootstrap the median of a vector.
}
