% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_cohort.R
\name{dev_cohort}
\alias{dev_cohort}
\title{Athlete Development Trend}
\usage{
dev_cohort(ath, top_k = 5, times = 10, ci_lims = c(min = 0.1, max = 0.9))
}
\arguments{
\item{ath}{character vector of athlete names}

\item{top_k}{integer; cohort includes athletes finishing in the `top_k`}

\item{times}{integer; cohort includes athletes finishing in the `top_k` at
least `times` times.}

\item{ci_lims}{named vector defining CI limits for estimating the median}
}
\description{
Plot athlete development relative to a cohort of top athletes at equivalent
ages.
}
