% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_cohort.R
\name{dev_cohort_dst_qgam}
\alias{dev_cohort_dst_qgam}
\title{importFrom qgam mqgam qgam}
\usage{
dev_cohort_dst_qgam(
  top_k = 10,
  times = 3,
  age = 30,
  q = c(0.1, 0.9),
  type = "fispoints"
)
}
\description{
importFrom qgam mqgam qgam
}
