% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_all_caps.R
\name{extract_all_caps}
\alias{extract_all_caps}
\title{Extract All Cap Words}
\usage{
extract_all_caps(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Helper function to extract only the portion of a string that is in all caps.
Useful for athlete name parsing since FIS displays names as: LAST_NAME First_Name.
}
