% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_races.R
\name{find_races}
\alias{find_races}
\title{Find Races}
\usage{
find_races(cur_date)
}
\arguments{
\item{date}{character; e.g. 'YYYY-MM-DD'}
}
\value{
A data frame with information on every race for the given date
}
\description{
Search for races on a specific date
}
\examples{
\dontrun{
find_races(cur_date = '2015-01-01')
}
}
