% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_skier.R
\name{find_skier}
\alias{find_skier}
\title{Simple Fuzzy Searching By Partial Name}
\usage{
find_skier(x)
}
\arguments{
\item{x}{character; a string you think should be in someone's name}
}
\value{
A data frame with unique individuals who's name contains
the given string
}
\description{
Since it is (mostly) easier to remember people's names than their
id numbers, this is a very simple function for looking up the "right"
spelling for an athlete's name.
}
\details{
Simply looks for names matching "\%string\%".
}
\examples{
\dontrun{
find_skier("OSGOOD")
find_skier("Bente")
}
}
