% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbump.R
\name{ggbump}
\alias{ggbump}
\title{importFrom tidyr gather}
\usage{
ggbump(
  grp_var,
  data,
  offset.x = 0.1,
  xlab = NULL,
  yrev = TRUE,
  title = NULL,
  subtitle = NULL
)
}
\description{
importFrom tidyr gather
}
