% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/head_to_head.R
\name{hth_dst}
\alias{hth_dst}
\title{Head-to-head Distance Plot & Data}
\usage{
hth_dst(ath1, ath2, races = c("maj_int", "fis"), measure = c("fispoints",
  "rank", "time", "pb", "mpb"), by_tech = FALSE, by_start = FALSE)
}
\arguments{
\item{ath1}{character; athlete of interest}

\item{ath2}{character vector of athletes to compare with}

\item{races}{character; one of "maj_int" or "fis"}

\item{measure}{character; one of "fispoints", "rank", "time", "pb" or "mpb".
"mpb" only is sensible if \code{races = "maj_int"}.}

\item{by_tech}{boolean; split trend lines by technique. Only one of
\code{by_tech} and \code{by_start} can by \code{TRUE}.}

\item{by_start}{boolean; split trend lines by race type. Only one of
\code{by_tech} and \code{by_start} can by \code{TRUE}.}
}
\description{
Head-to-head Distance Plot & Data
}

