% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_to_head_lmer.R
\name{hth_lmer}
\alias{hth_lmer}
\title{Assess Head-to-Head Performance Using lmer Model}
\usage{
hth_lmer(
  ath_names,
  event_id,
  num_opp = Inf,
  cutoff = 365 * 5,
  min_encounters = 1,
  measure = c("rank", "fispoints", "pb"),
  events = c("all", "maj_int"),
  restrict_by = NULL,
  race_title = ""
)
}
\arguments{
\item{ath_names}{character vector of athlete names}

\item{event_id}{integer; id of the race to assess performance of}

\item{num_opp}{integer; number of opponents to consider in each race, if `Inf`
use all opponents}

\item{cutoff}{integer; size of time window in days to collect prior races from}

\item{min_encounters}{integer; only include opponents who have faced an athlete
at least this many times}

\item{measure}{character; one of "rank", "fispints" or "pb" (percent back)}

\item{restrict_by}{currently unused}

\item{race_title}{character; race title to add to the plot}
}
\description{
Fit an `lmer()` model to prior head-to-head results for a set of skiers
in order to plot expected vs observed performances. This is attempting to
serve the same general purpose as `head_to_head_race()` but using a more
"sophisticated" method.
}
