% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_to_head_race.R
\name{hth_race}
\alias{hth_race}
\title{Head-to-Head Against Skiers in a Specific Race}
\usage{
hth_race(ath1_name, race_id, num_opp = Inf, cutoff = 365 * 5,
  min_encounters = 1, measure = c("rank", "fispoints", "pb"),
  restrict_by = NULL)
}
\arguments{
\item{ath1_name}{character}

\item{race_id}{integer}

\item{num_opp}{integer; limit comparison to opponents in the top \code{num_opp}.
Defaults to \code{Inf} meaning look at all opponents.}

\item{cutoff}{integer; length of time in days to look at historical results}

\item{min_encounters}{integer; minimum number of time \code{ath1_name} must have
faced an opponent for the opponent to be included in the comparison}

\item{measure}{character; one of \code{"rank"}, \code{"fispoints"} or \code{"pb"}}

\item{restrict_by}{currently unused}
}
\description{
Compare results of \code{ath1_name} in a specific race (\code{race_id})
to the historical record of how \code{ath1_name} has performed against the
particular opponents in that race. Serves as a slightly more "objective" way
to judge whether \code{ath1_name} had a good, bad or average race.
}
\examples{
\dontrun{
x <- hth_race('DIGGINS Jessica',9232,num_opp = 30,min_encounters = 3,measure = 'pb')
}
}

