% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_to_head.R
\name{hth_spr}
\alias{hth_spr}
\title{Head-to-head Sprint Plot & Data}
\usage{
hth_spr(ath1, ath2, races = c("maj_int", "fis"), measure = c("fispoints",
  "rank"), by_tech = FALSE)
}
\arguments{
\item{ath1}{character; athlete of interest}

\item{ath2}{character vector of athletes to compare with}

\item{races}{character; one of "maj_int" or "fis"}

\item{measure}{character; one of "fispoints" or "rank"}

\item{by_tech}{boolean; split trend lines by technique.}
}
\description{
Head-to-head Sprint Plot & Data
}

