% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_urls.R
\name{insert_race_url}
\alias{insert_race_url}
\title{Insert Race URLs}
\usage{
insert_race_url(raceid, url1 = NA, url2 = NA)
}
\arguments{
\item{raceid}{integer race id}

\item{url1}{character}

\item{url2}{character}
}
\description{
Only available to those with INSERT priviledges
}
