% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_best_n.R
\name{lag_best_avg}
\alias{lag_best_avg}
\title{Lagged Average of Top Results}
\usage{
lag_best_avg(x, dates, n = 5L, window = 365L, adj = c(1.4, 1.3, 1.2, 1.1,
  1))
}
\arguments{
\item{x}{numeric vector of values}

\item{dates}{dates, assumed to be character in YYYY-MM-DD format}

\item{n}{integer}

\item{window}{integer; previous number of days to examine}

\item{adj}{vector of adjustments for when there are fewer than \code{n}
values.}
}
\description{
FIS points are typically the average of the best 5 results over the prior
calendar year. This is a fast, Rcpp way to replicate this calculation.
}
