% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_best_n.R
\name{lag_mad}
\alias{lag_mad}
\title{Lagged Median All Results}
\usage{
lag_mad(x, dates, window = 365L)
}
\arguments{
\item{x}{numeric vector of values}

\item{dates}{dates, assumed to be character in YYYY-MM-DD format}

\item{window}{integer; previous number of days to examine}

\item{n}{integer}

\item{adj}{vector of adjustments for when there are fewer than \code{n}
values.}
}
\description{
Median result over the past calendar window in days.
}
