% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_best_n.R
\name{lag_median}
\alias{lag_median}
\title{Lagged Median All Results}
\usage{
lag_median(x, dates, window = 365L, adj = c(1.4, 1.3, 1.2, 1.1, 1))
}
\arguments{
\item{x}{numeric vector of values}

\item{dates}{dates, assumed to be character in YYYY-MM-DD format}

\item{window}{integer; previous number of days to examine}

\item{adj}{vector of adjustments for when there are fewer than \code{n}
values.}

\item{n}{integer}
}
\description{
Median result over the past calendar window in days.
}
