% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nation_trend.R
\name{nation_dst_trend}
\alias{nation_dst_trend}
\title{Performance Trends By Nation}
\usage{
nation_dst_trend(
  .nation,
  levels = c(Wins = 1, Podiums = 3, Top10 = 10, Top30 = 30)
)
}
\arguments{
\item{.nation}{character vector of nation codes}
}
\value{
A named list with components:
\enumerate{
 \item \code{plot} - ggplot2 plot object
 \item \code{nation_data} - raw data for each nation
 \item \code{nation_summary} - summarised data by nation
 \item \code{race_data} - race counts used for plotting
}
}
\description{
Only considers results from major international races and
summarises them by the number of race results at a particular
level (wins, podiums, etc.) per race.
}
\examples{
\dontrun{
library(ggplot2)
p <- nation_trend(nations = c('USA','CAN','RUS','SWE'),
                  .gender = 'Men',
                  .type = 'Distance')
print(p$plot)
}
}
