% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nation_trend.R
\name{nation_trend}
\alias{nation_trend}
\title{National Level Performance Trends}
\usage{
nation_trend(nations, race_gender = c("Men", "Women"),
  race_type = c("Distance", "Sprint"))
}
\arguments{
\item{nations}{character vector of nations}

\item{race_gender}{character; one of "Men" or "Women"}

\item{race_type}{character; one of "Distance" or "Sprint"}
}
\description{
Only look at major international results.
}

