% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_urls.R
\name{open_race_url}
\alias{open_race_url}
\title{Open Race URL in Browser}
\usage{
open_race_url(raceid)
}
\arguments{
\item{raceid}{integer}
}
\description{
Open results for a race in a web browser.
}
\examples{
\dontrun{
open_race_url(raceid = 7902)
}
}
\seealso{
Other race.info.functions: \code{\link{update_race_url}}
}

