% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_best_n.R
\name{pts_thresh}
\alias{pts_thresh}
\title{FIS Point Threshold Necessary to Raise/Lower Athlete's Point}
\usage{
pts_thresh(x, dates, window = 365L)
}
\arguments{
\item{x}{numeric vector of FIS points}

\item{dates}{character vector of dates, must be in YYYY-MM-DD format}

\item{window}{integer number of days}
}
\description{
Calculate the point threshold at any given point in time at which an athlete
would be able to lower their carried FIS points.
}
