% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/race_snapshot.R
\name{race_snapshot_spr}
\alias{race_snapshot_spr}
\title{Sprint Race Snapshot}
\usage{
race_snapshot_spr(race_id, cutoff = 365 * 4, reduced = TRUE)
}
\arguments{
\item{race_id}{integer race id}

\item{cutoff}{integer number of days in past to gether data from}

\item{reduced}{boolean; only examine top 30 plus North Americans}
}
\description{
Race snapshot graphs for major international
sprint races.
}

