% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_start_list.R
\name{read_start_list}
\alias{read_start_list}
\title{Scrape Race Start Lists}
\usage{
read_start_list(url, type = c("sprint", "distance"))
}
\description{
Simple function to grab start lists from FIS website.
}
