% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_events.R
\name{recent_events}
\alias{recent_events}
\title{Recent Races}
\usage{
recent_events(cur_date = as.character(Sys.Date() - 14))
}
\arguments{
\item{cur_date}{character; a date in the format "YYYY-MM-DD"}
}
\value{
A data frame with one row for each race listing various details
}
\description{
Utility function for listing the most recent
events in the database since a given date. Used mainly
to help track what new races need to be entered.
}
\examples{
\dontrun{
x <- recent_events("2016-02-01")
head(x)
}
}
