% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_races.R
\name{recent_races}
\alias{recent_races}
\title{Recent Races}
\usage{
recent_races(cur_date)
}
\arguments{
\item{cur_date}{character; a date in the format "YYYY-MM-DD"}
}
\value{
A data frame with one row for each race listing various details
}
\description{
Utility function for listing the most recent
races in the database since a given date. Used mainly
to help track what new races need to be entered.
}
\examples{
\dontrun{
x <- recent_races("2016-02-01")
head(x)
}
}
