% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_to_date.R
\name{season_to_date}
\alias{season_to_date}
\title{Convert season string to date}
\usage{
season_to_date(x)
}
\arguments{
\item{x}{character; season to convert}
}
\description{
Converts season string like \code{2010-2011} to a date string like
\code{2011-01-01}.
}
