% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_all_caps.R
\name{shorten_names}
\alias{shorten_names}
\title{Shorten Skier Names}
\usage{
shorten_names(x)
}
\arguments{
\item{x}{character vector of skier names}
}
\description{
Grabs the all caps portion and applies title case
}
