% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr_plot.R
\name{spr_plot}
\alias{spr_plot}
\title{Sprint Results Plot}
\usage{
spr_plot(ath_names, by_tech = FALSE)
}
\arguments{
\item{ath_names}{character vector of athlete names}

\item{by_tech}{boolean; if \code{TRUE} split graph by technique}
}
\value{
A named list with components:
\enumerate{
 \item \code{plot} - a ggplot2 plot object
 \item \code{ath_data} - raw athlete data
}
}
\description{
Bar plot by season summarising major international sprint
race performance by counting how often an athlete reached the
quaterfinals, semifinals, etc.
}
\examples{
\dontrun{
library(ggplot2)
p <- spr_plot(ath_names = c('NEWELL Andrew','HAMILTON Simeon'))
print(p$plot)
}
}
