% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_perc_back.R
\name{standardize_mpb}
\alias{standardize_mpb}
\title{Standardize Median Percent Back}
\usage{
standardize_mpb(race_data)
}
\arguments{
\item{race_data}{data frame}
}
\description{
Standardize median percent back by
season, gender and race type
}

