% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{statskier_connect}
\alias{statskier_connect}
\title{Create Connection to StatSkier DB}
\usage{
statskier_connect()
}
\value{
A database connection object
}
\description{
Call this function with no arguments to create a hidden object in the
global environment that acts as the db connection. This function expects
all the connection information to be stored in `options()$statskier_pg` as
a list with elements host, dbname, user, password, port and sslmode.
}
\details{
The hidden db connection object is called \code{..statskier_pg_con..}.
}
