% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_urls.R
\name{update_race_url}
\alias{update_race_url}
\title{Update Race URLs}
\usage{
update_race_url(raceid, url1 = NA, url2 = NA)
}
\arguments{
\item{raceid}{integer race id}

\item{url1}{character}

\item{url2}{character}
}
\description{
Update Race URLs
}
\seealso{
Other race.info.functions: \code{\link{insert_race_url}},
  \code{\link{open_race_url}}
}

