% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_data.R
\name{dst_elo_data}
\alias{dst_elo_data}
\title{Generate ELO Distance Input Data}
\usage{
dst_elo_data(default_rating = 1300)
}
\arguments{
\item{default_rating}{skiers default rating}
}
\description{
Generate ELO Distance Input Data
}

