% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/median_perc_back.R
\name{mpb}
\alias{mpb}
\title{Median Percent Back}
\usage{
mpb(race_data)
}
\arguments{
\item{race_data}{data frame with race results}
}
\description{
Calculate the percent back from the median skier
}

