% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_urls.R
\name{open_race_url}
\alias{open_race_url}
\title{Show race in browser}
\usage{
open_race_url(raceid)
}
\arguments{
\item{raceid}{integer}
}
\description{
Open results for a race in a web browser.
}
\seealso{
Other race.info.functions: \code{\link{insert_race_url}},
  \code{\link{update_race_url}}
}

