% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_snapshot.R
\name{race_snapshot_dst}
\alias{race_snapshot_dst}
\alias{race_snapshot_spr}
\title{Race Snapshots}
\usage{
race_snapshot_dst(event_id, title = "", cutoff = 365 * 4, reduced = TRUE)

race_snapshot_spr(event_id, title = "", cutoff = 365 * 4, reduced = TRUE)
}
\arguments{
\item{title}{race title}

\item{cutoff}{integer number of days in past to gether data from}

\item{reduced}{boolean; if \code{TRUE} only show top 30 plus North Americans}

\item{race_id}{integer race id}
}
\value{
A named list with components:
\enumerate{
 \item \code{plot} - ggplot2 plot object
 \item \code{cur_race} - raw results for race in question
 \item \code{ath_min} - data for athletes with too few races to plot range bars
 \item \code{ath_bars} - data for athlete range bars
 \item \code{block} - data frame for color blocks in plot
}
}
\description{
Distance and sprint race snapshot graphs for major international
distance events.
}
\examples{
\dontrun{
library(ggplot2)
p <- race_snapshot_dst(race_id = 7902)
print(p$plot)
}
}
