% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_races.R
\name{recent_races}
\alias{recent_races}
\title{List most recent XC races in db}
\usage{
recent_races(cur_date)
}
\arguments{
\item{cur_date}{character; a date}
}
\description{
List most recent XC races in db
}

