% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/median_perc_back.R
\name{standardize_mpb}
\alias{standardize_mpb}
\title{Standardize Median Percent Back}
\usage{
standardize_mpb(race_data)
}
\arguments{
\item{race_data}{}
}
\description{
Standardize by season, gender and race type
}

