% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wjc_u23_hist.R
\name{wjc_u23_plot}
\alias{wjc_u23_plot}
\title{WJC & U23 History Plots}
\usage{
wjc_u23_plot(nations, races = c("wjc", "wu23"))
}
\arguments{
\item{nations}{character vector of nation codes, eg "USA", "CAN", etc.}

\item{races}{character; one of "wjc" or "wu23"}
}
\value{
A named list with components:
\enumerate{
 \item \code{plots} - a named list of plots, one for each nation
 \item \code{data} - raw data used for plots
}
}
\description{
Results trends for nations at either World Juniors or
U23s, by gender and race type.
}
\examples{
\dontrun{
p <- wjc_u23_plot(nations = c('USA','CAN'),races = 'wjc')
print(p$plots$USA)
print(p$plots$CAN)
}
}
