% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wjc_u23_hist.R
\name{wjc_u23_plot}
\alias{wjc_u23_plot}
\title{WJC & U23 History Plots & Data}
\usage{
wjc_u23_plot(nations, races = c("WJC", "U23"))
}
\arguments{
\item{nations}{character vector of nations}

\item{races}{character; one of "WJC" or "U23"}
}
\description{
WJC & U23 History Plots & Data
}

